% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addspaces.R
\name{addspaces}
\alias{addspaces}
\title{addspaces}
\usage{
addspaces(string, L)
}
\arguments{
\item{string}{phrase}

\item{L}{number of letters before a new space}
}
\value{
a string
}
\description{
Internal function to add spaces between n letters in a string.
This might be helpful when printing long sequences of ciphered text,
}
\examples{
addspaces("abcdefghijklm", 5)

}
