% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\name{affine}
\alias{affine}
\title{affine}
\usage{
affine(word, a, b, encrypt = TRUE)
}
\arguments{
\item{word}{Word or phrase to be encrypted}

\item{a}{First parameter. This value and 26 must be coprime}

\item{b}{Second parameter. Magnitude of the shift}

\item{encrypt}{If `TRUE` (default), the program ciphers the input word, If `FALSE`, the program decrypts it.}
}
\value{
a string
}
\description{
The affine cipher is a monoalphabetic substitutoin cipher,
where each letter is enciphered with the function (ax+b) mod 26 
(26 is the number of letters in the alphabet)
}
\examples{
affine("Hello", 1, -1)

}
\references{
https://en.wikipedia.org/wiki/Affine_cipher
}
