% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atbash.R
\name{atbash}
\alias{atbash}
\title{atbash}
\usage{
atbash(word)
}
\arguments{
\item{word}{Word or phrase to be encrypted}
}
\value{
a string
}
\description{
The Atbash cipher is a type of monoalphabetic cipher which takes
 the alphabet and maps it to its reverse. 
 It is a particular case of the affine cipher, with `a`=`b`= (`m`-1). As `m` is the number
 of letters and is equal to 26, it means that `a` = `b` = 25.
 Encrypting and decrypting are not separate for this cipher.
}
\examples{
atbash("abcxyz")

}
\references{
https://en.wikipedia.org/wiki/Atbash
}
