% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caesar.R
\name{caesar}
\alias{caesar}
\title{caesar}
\usage{
caesar(word, key = 1, encrypt = TRUE)
}
\arguments{
\item{word}{Word or phrase to be encrypted}

\item{key}{numeric key}

\item{encrypt}{If `TRUE` (default), the program ciphers the input word, If `FALSE`, the program decrypts it.}
}
\value{
a string
}
\description{
caesar encryption
}
\examples{
caesar("Hello", 1)

}
