% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullcipher.R
\name{nullcipher}
\alias{nullcipher}
\title{nullcipher}
\usage{
nullcipher(phrase, index, decrypt = TRUE)
}
\arguments{
\item{phrase}{Word or phrase to be decrypted}

\item{index}{letter of interest for each word in the phrase. Also a pattern vector can be entered.}

\item{decrypt}{Only Decryption is possible for now, but will be updated in the future}
}
\value{
a string
}
\description{
A null cipher is an encryption method where the plaintext is mixed with a
 large amount of non-cipher material (decoy).
}
\examples{
nullcipher("handy set false posts", c(1,2,3))

}
\references{
https://en.wikipedia.org/wiki/Null_cipher
}
