% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepCyp.w.R
\name{PrepCyp.w}
\alias{PrepCyp.w}
\title{PrepCyp.w}
\usage{
PrepCyp.w(word)
}
\arguments{
\item{word}{Word or phrase to be preprocessed}
}
\value{
a string
}
\description{
Internal function for preparing the input word (enforcing lowercase, ASCII standard, and no blanks)
}
\examples{
PrepCyp.w("HèllO Wòrld")

}
