% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cipher.R
\name{print.cipher}
\alias{print.cipher}
\title{print.cipher}
\usage{
\method{print}{cipher}(x, ...)
}
\arguments{
\item{x}{previous object}

\item{...}{other arguments}
}
\value{
nothing
}
\description{
internal function with print definitions for custom cipher class
}
