% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/railfence.R
\name{railfence}
\alias{railfence}
\title{railfence}
\usage{
railfence(word, key)
}
\arguments{
\item{word}{Word or phrase to be encrypted}

\item{key}{numeric key (number of rails)}
}
\value{
a string
}
\description{
rail fence cipher
}
\examples{
railfence('we are discovered flee at once',3)

}
