% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_substitution.R
\name{simple_substitution}
\alias{simple_substitution}
\title{simple_substitution}
\usage{
simple_substitution(word, key = "", seed = sample(1:1000, 1))
}
\arguments{
\item{word}{Word or phrase to be encrypted}

\item{key}{Word to be used as key for the encryption. If not provided, a random shuffle is performed}

\item{seed}{Seed for reproducibility of the encryption if key is not provided}
}
\value{
a list with custom class "cipher", which modifies the printing defaults.
        The list contains the initial phrase (initial), the ciphered output (encrypted),
         and the alphabet order (keyalphabet)
}
\description{
simple substitution cipher. Each letter is monoalphabetically associated with a different one used for the encryption.
}
\examples{
simple_substitution("hello world", seed = 1234)
simple_substitution("hello world", key = "zebras")

}
