% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecolumn.R
\name{singlecolumn}
\alias{singlecolumn}
\title{singlecolumn}
\usage{
singlecolumn(word, key, rm.blanks = TRUE)
}
\arguments{
\item{word}{Word or phrase to be encrypted}

\item{key}{word key: for example, the key "bcea" suggests that the column order is "2-3-4-1"}

\item{rm.blanks}{Should spaces between words be removed? By default set to `TRUE`}
}
\value{
a string
}
\description{
In a columnar transposition cipher, the message is written out in rows of a fixed length,
 and then read out again column by column. The order of the column follows the alphabetical order of the letters present in the key
}
\examples{
singlecolumn("This is wikipedia", "cipher")

}
\references{
https://www.geeksforgeeks.org/columnar-transposition-cipher/
}
