% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trancipher_wordkey.R
\name{trancipher_wordkey}
\alias{trancipher_wordkey}
\title{trancipher_wordkey}
\usage{
trancipher_wordkey(word, key)
}
\arguments{
\item{word}{Word or phrase to be encrypted}

\item{key}{numeric key}
}
\value{
a string
}
\description{
transposition ciphertext with a word as key
}
\examples{
trancipher_wordkey("This is wikipedia", "cipher")

}
