% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caesar.R
\name{caesar}
\alias{caesar}
\title{caesar}
\usage{
caesar(word, key, mode = "encrypt")
}
\arguments{
\item{word}{Word or phrase to be encrypted}

\item{key}{numeric key}

\item{mode}{One of "decrypt" or "encrypt". Sets whether the program encrypts or decrypts the input word}
}
\value{
a string
}
\description{
caesar encryption
}
\examples{
caesar("Hello", 1, "encrypt")

}
