% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playfair.R
\name{playfair}
\alias{playfair}
\title{playfair}
\usage{
playfair(word, key, added_letter = "x", decrypt = FALSE)
}
\arguments{
\item{word}{Word or phrase to be encrypted}

\item{key}{Word for creating the modified Polybius square}

\item{added_letter}{Letter to be added in case 2 letters in a pair are identical; usually "x" is used}

\item{decrypt}{If `FALSE` (default), the program ciphers the input word, If `TRUE`, the program decrypts it.}
}
\value{
a string
}
\description{
playfair encryption
}
\examples{
playfair("instruments", "monarchy", added_letter = "z")

}
\references{
https://en.wikipedia.org/wiki/Playfair_cipher
}
