% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scytale.R
\name{scytale}
\alias{scytale}
\title{scytale}
\usage{
scytale(word, key = 3, encrypt = TRUE)
}
\arguments{
\item{word}{Word or phrase to be encrypted or decrypted}

\item{key}{Number of turns of the band}

\item{encrypt}{If `TRUE` (default), the program ciphers the input word, If `FALSE`, the program decrypts it.}
}
\value{
a string
}
\description{
The Scytale is a transposition cipher
 The diameter of the Scytale (the number of turns) can be regarded as the key of the cipher.
}
\examples{
scytale('we are discovered flee at once',3)

}
\references{
https://en.wikipedia.org/wiki/Scytale
}
