% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecolumn.R
\name{singlecolumn}
\alias{singlecolumn}
\title{singlecolumn}
\usage{
singlecolumn(word, key, rm.blanks = TRUE)
}
\arguments{
\item{word}{Word or phrase to be encrypted}

\item{key}{numeric key}

\item{rm.blanks}{Should spaces between words be removed? By default set to TRUE}
}
\value{
a string
}
\description{
transposition ciphertext with a word as key
}
\examples{
singlecolumn("This is wikipedia", "cipher")

}
