% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenere.R
\name{vigenere}
\alias{vigenere}
\title{vigenere}
\usage{
vigenere(word, key, encrypt = TRUE)
}
\arguments{
\item{word}{Word or phrase to be encrypted}

\item{key}{character key}

\item{encrypt}{If `TRUE` (default), the program ciphers the input word, If `FALSE`, the program decrypts it.}
}
\value{
a string
}
\description{
Vigenère cipher is a method of encrypting alphabetic text where each letter of
 the plaintext is encoded with a different Caesar cipher, whose increment is determined by the corresponding letter the key
}
\examples{
vigenere("hello world", "opla")

}
\references{
https://en.wikipedia.org/wiki/Vigen%C3%A8re_cipher
}
